requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
fastembed==0.2.2
pydantic==2.5.3
python-dotenv==1.0.0
qdrant-client==1.7.0
requests==2.31.0
